requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
python-pushover==0.4
requests==2.28.0
urllib3==1.26.9
